+++ flutter_extensions/lib/src/widgets/fixed_indexed_stack.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-/// Build the child only when the index is selected
+// Build the child only when the index is selected
 class FixedIndexedStack extends StatefulWidget {
   /// see [IndexedStack.index]
   final int index;
@@ -17,47 +17,53 @@ class FixedIndexedStack extends StatefulWidget {
   /// see [MultiChildRenderObjectWidget.children]
   final List<Widget> children;
 
-  const FixedIndexedStack({
+  FixedIndexedStack({
     Key? key,
     this.index = 0,
     this.alignment = AlignmentDirectional.topStart,
     this.textDirection,
     this.sizing = StackFit.loose,
     this.children = const <Widget>[],
-  }) : super(key: key);
+  })  : assert(children.every((child) => child.key != null)),
+        super(key: key);
 
   @override
   _FixedIndexedStackState createState() => _FixedIndexedStackState();
 }
 
 class _FixedIndexedStackState extends State<FixedIndexedStack> {
-  late List<Widget> _children;
+  late Set<Key> _indexedChildren;
 
   @override
   void initState() {
     super.initState();
-    _children = List.generate(widget.children.length, (index) => const SizedBox.shrink());
-    _children[widget.index] = widget.children[widget.index];
+    _indexedChildren = {};
   }
 
   @override
   void didUpdateWidget(covariant FixedIndexedStack oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (widget.children.length != oldWidget.children.length) {
-      _children.length = widget.children.length;
+
+    _indexedChildren.removeWhere((key) => !widget.children.any((child) => child.key == key));
+
+    if (widget.index != oldWidget.index) {
+      final child = widget.children[widget.index];
+      _indexedChildren.add(child.key!);
     }
-    // add or update child when index change
-    _children[widget.index] = widget.children[widget.index];
   }
 
   @override
   Widget build(BuildContext context) {
+    final children = widget.children.map((child) {
+      return _indexedChildren.contains(child.key) ? child : const SizedBox.shrink();
+    }).toList();
+
     return IndexedStack(
       index: widget.index,
       alignment: widget.alignment,
       textDirection: widget.textDirection,
       sizing: widget.sizing,
-      children: _children,
+      children: children,
     );
   }
 }