+++ pure_extensions/lib/src/built/collections.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import 'package:built_collection/built_collection.dart';
 import 'package:collection/collection.dart';
+import 'package:pure_extensions/src/dart/errors.dart';
 
 extension BuiltListExt<T> on BuiltList<T> {
   T random({Random? random}) {
@@ -48,10 +49,13 @@ extension SetBuilderExt<T> on SetBuilder<T> {
 }
 
 extension BuiltMapExt<K, V> on BuiltMap<K, V> {
-  /// Returns a value
-  V get(K key) => this[key]!;
+  /// Returns the value [key].
+  V get(K key) {
+    final value = this[key];
+    return value is V ? value : throw KeyError(key, this);
+  }
 
-  /// Returns a value or null if it not exist
+  /// Returns the value [key] or null if it not exist.
   V? getOrNull(K key) => this[key];
 
   /// returns a List from entries